Adjust Kaz holidays to master and update tests
