+++ holidays/countries/kazakhstan.py
@@ -4,7 +4,7 @@ from datetime import date
 from dateutil.relativedelta import relativedelta as rd
 from holidays.constants import JAN, MAR, MAY, JUL, AUG, DEC
 from holidays.holiday_base import HolidayBase
-from holidays.utils import get_gre_date
+from holidays.utils import islamic_to_gre
 
 
 class Kazakhstan(HolidayBase):
@@ -51,7 +51,7 @@ class Kazakhstan(HolidayBase):
         self[date(year, JUL, 6)] = "Capital City Day"
 
         # Kurban Ait
-        for hol_date in get_gre_date(year, 12, 10):
+        for hol_date in islamic_to_gre(year, 12, 10):
             self[hol_date] = "Kurban Ait"
 
         # Constitution Day
